requirements.txt
Flask==2.0.1
PyJWT==2.1.0
requests==2.25.1
flask_swagger_ui
unittest